refactor: Deprecate RagAgent and Model in favor of versatile Agent
