+++ rig-core/examples/simple_model.rs
@@ -5,7 +5,7 @@ async fn main() {
     // Create OpenAI client and model
     let openai_client = openai::Client::from_env();
 
-    let gpt4 = openai_client.model("gpt-4").build();
+    let gpt4 = openai_client.agent("gpt-4").build();
 
     // Prompt the model and print its response
     let response = gpt4
