+++ rig-core/src/model.rs
@@ -47,56 +47,11 @@
 //!     .await
 //!     .expect("Failed to send completion request");
 //! ```
-use crate::completion::{
-    Chat, Completion, CompletionError, CompletionModel, CompletionRequestBuilder,
-    CompletionResponse, Message, ModelChoice, Prompt, PromptError,
-};
-
-/// A model that can be used to prompt completions from a completion model.
-/// This is the simplest building block for creating an LLM powered application.
-pub struct Model<M: CompletionModel> {
-    /// Completion model (e.g.: OpenAI's `gpt-3.5-turbo-1106`, Cohere's `command-r`)
-    model: M,
-    /// Temperature of the model
-    temperature: Option<f64>,
-}
-
-impl<M: CompletionModel> Completion<M> for Model<M> {
-    async fn completion(
-        &self,
-        prompt: &str,
-        chat_history: Vec<Message>,
-    ) -> Result<CompletionRequestBuilder<M>, CompletionError> {
-        Ok(self
-            .model
-            .completion_request(prompt)
-            .messages(chat_history)
-            .temperature_opt(self.temperature))
-    }
-}
-
-impl<M: CompletionModel> Prompt for Model<M> {
-    async fn prompt(&self, prompt: &str) -> Result<String, PromptError> {
-        self.chat(prompt, vec![]).await
-    }
-}
-
-impl<M: CompletionModel> Chat for Model<M> {
-    async fn chat(&self, prompt: &str, chat_history: Vec<Message>) -> Result<String, PromptError> {
-        match self.completion(prompt, chat_history).await?.send().await? {
-            CompletionResponse {
-                choice: ModelChoice::Message(message),
-                ..
-            } => Ok(message),
-            CompletionResponse {
-                choice: ModelChoice::ToolCall(toolname, _),
-                ..
-            } => Err(PromptError::ToolError(
-                crate::tool::ToolSetError::ToolNotFoundError(toolname),
-            )),
-        }
-    }
-}
+#[deprecated(
+    since = "0.2.0",
+    note = "Please use the `Agent` type directly instead of the `Model` type."
+)]
+pub type Model<M> = crate::agent::Agent<M>;
 
 /// A builder for creating a model
 ///
@@ -113,37 +68,9 @@ impl<M: CompletionModel> Chat for Model<M> {
 ///     .temperature(0.8)
 ///     .build();
 /// ```
-pub struct ModelBuilder<M: CompletionModel> {
-    model: M,
-    temperature: Option<f64>,
-}
 
-impl<M: CompletionModel> ModelBuilder<M> {
-    /// Create a new model builder
-    pub fn new(model: M) -> Self {
-        Self {
-            model,
-            temperature: None,
-        }
-    }
-
-    /// Set the temperature of the model
-    pub fn temperature(mut self, temperature: f64) -> Self {
-        self.temperature = Some(temperature);
-        self
-    }
-
-    /// Set the temperature of the model (set to `None` to use the default temperature of the model)
-    pub fn temperature_opt(mut self, temperature: Option<f64>) -> Self {
-        self.temperature = temperature;
-        self
-    }
-
-    /// Build the model
-    pub fn build(self) -> Model<M> {
-        Model {
-            model: self.model,
-            temperature: self.temperature,
-        }
-    }
-}
+#[deprecated(
+    since = "0.2.0",
+    note = "Please use the `AgentBuilder` type directly instead of the `ModelBuilder` type."
+)]
+pub type ModelBuilder<M> = crate::agent::AgentBuilder<M>;
