+++ rig-core/src/providers/perplexity.rs
@@ -14,8 +14,6 @@ use crate::{
     completion::{self, CompletionError},
     extractor::ExtractorBuilder,
     json_utils,
-    model::ModelBuilder,
-    rag::RagAgentBuilder,
 };
 
 use schemars::JsonSchema;
@@ -73,8 +71,12 @@ impl Client {
         CompletionModel::new(self.clone(), model)
     }
 
-    pub fn model(&self, model: &str) -> ModelBuilder<CompletionModel> {
-        ModelBuilder::new(self.completion_model(model))
+    #[deprecated(
+        since = "0.2.0",
+        note = "Please use the `agent` method instead of the `model` method."
+    )]
+    pub fn model(&self, model: &str) -> AgentBuilder<CompletionModel> {
+        AgentBuilder::new(self.completion_model(model))
     }
 
     pub fn agent(&self, model: &str) -> AgentBuilder<CompletionModel> {
@@ -88,12 +90,20 @@ impl Client {
         ExtractorBuilder::new(self.completion_model(model))
     }
 
-    pub fn rag_agent(&self, model: &str) -> RagAgentBuilder<CompletionModel> {
-        RagAgentBuilder::new(self.completion_model(model))
+    #[deprecated(
+        since = "0.2.0",
+        note = "Please use the `agent` method instead of the `rag_agent` method."
+    )]
+    pub fn rag_agent(&self, model: &str) -> AgentBuilder<CompletionModel> {
+        AgentBuilder::new(self.completion_model(model))
     }
 
-    pub fn context_rag_agent(&self, model: &str) -> RagAgentBuilder<CompletionModel> {
-        RagAgentBuilder::new(self.completion_model(model))
+    #[deprecated(
+        since = "0.2.0",
+        note = "Please use the `agent` method instead of the `context_rag_agent` method."
+    )]
+    pub fn context_rag_agent(&self, model: &str) -> AgentBuilder<CompletionModel> {
+        AgentBuilder::new(self.completion_model(model))
     }
 }
